経験値とコインの紐づけの修正

+++ Assets/Scripts/WholeScripts/GrobalInfo.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class GrobalInfo : MonoBehaviour
 {
-    public static int level = 0;
+    public static float level = 0;
     public static int coin = 10000;
     public static int life = 5;
